GITLAB-328 Dealing with CassandraACLMapper is the job of CassandraMailboxMapper, not of its DAO
