+++ mailbox/cassandra/src/main/java/org/apache/james/mailbox/cassandra/mail/CassandraMailboxMapper.java
@@ -107,9 +107,25 @@ public class CassandraMailboxMapper implements MailboxMapper {
     @Override
     public Mailbox findMailboxById(MailboxId id) throws MailboxException {
         CassandraId mailboxId = (CassandraId) id;
-        return mailboxDAO.retrieveMailbox(mailboxId)
+        return retrieveMailbox(mailboxId)
             .join()
             .orElseThrow(() -> new MailboxNotFoundException(id.serialize()));
+    }
+
+    private CompletableFuture<Optional<SimpleMailbox>> retrieveMailbox(CassandraId mailboxId) {
+        CompletableFuture<MailboxACL> aclCompletableFuture = cassandraACLMapper.getACL(mailboxId);
+
+        CompletableFuture<Optional<SimpleMailbox>> simpleMailboxFuture = mailboxDAO.retrieveMailbox(mailboxId);
+
+        return CompletableFutureUtil.combine(
+            aclCompletableFuture,
+            simpleMailboxFuture,
+            this::addAcl);
+    }
+
+    private Optional<SimpleMailbox> addAcl(MailboxACL acl, Optional<SimpleMailbox> mailboxOptional) {
+        mailboxOptional.ifPresent(mailbox -> mailbox.setACL(acl));
+        return mailboxOptional;
     }
 
     @Override
@@ -124,7 +140,7 @@ public class CassandraMailboxMapper implements MailboxMapper {
     }
 
     private CompletableFuture<Optional<SimpleMailbox>> retrieveMailbox(CassandraMailboxPathDAO.CassandraIdAndPath idAndPath) {
-        return mailboxDAO.retrieveMailbox(idAndPath.getCassandraId())
+        return retrieveMailbox(idAndPath.getCassandraId())
             .thenApply(optional -> OptionalUtils.ifEmpty(optional,
                 () -> LOGGER.warn("Could not retrieve mailbox {} with path {} in mailbox table.", idAndPath.getCassandraId(), idAndPath.getMailboxPath())));
     }
@@ -150,7 +166,7 @@ public class CassandraMailboxMapper implements MailboxMapper {
     private CompletableFuture<Boolean> trySave(SimpleMailbox cassandraMailbox, CassandraId cassandraId) {
         return mailboxPathDAO.save(cassandraMailbox.generateAssociatedPath(), cassandraId)
             .thenCompose(CompletableFutureUtil.composeIfTrue(
-                () -> mailboxDAO.retrieveMailbox(cassandraId)
+                () -> retrieveMailbox(cassandraId)
                     .thenCompose(optional -> CompletableFuture
                         .allOf(optional
                                 .map(storedMailbox -> mailboxPathDAO.delete(storedMailbox.generateAssociatedPath()))
@@ -189,6 +205,7 @@ public class CassandraMailboxMapper implements MailboxMapper {
     @Override
     public List<Mailbox> list() throws MailboxException {
         return mailboxDAO.retrieveAllMailboxes()
+            .thenComposeOnAll(this::toMailboxWithAclFuture)
             .join()
             .collect(Guavate.toImmutableList());
     }
@@ -229,6 +246,15 @@ public class CassandraMailboxMapper implements MailboxMapper {
         } else {
             return Pattern.quote((String) token);
         }
+    }
+
+    private CompletableFuture<SimpleMailbox> toMailboxWithAclFuture(SimpleMailbox mailbox) {
+        CassandraId cassandraId = (CassandraId) mailbox.getMailboxId();
+        return cassandraACLMapper.getACL(cassandraId)
+            .thenApply(acl -> {
+                mailbox.setACL(acl);
+                return mailbox;
+            });
     }
 
 }
